+++ .storybook/main.ts
@@ -1,0 +1,29 @@
+import type { StorybookConfig } from '@storybook/react-vite'
+import { mergeConfig } from 'vite'
+import { resolve } from 'path'
+
+const config: StorybookConfig = {
+  stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
+  addons: [
+    '@storybook/addon-links',
+    '@storybook/addon-essentials',
+    '@storybook/addon-interactions',
+  ],
+  framework: {
+    name: '@storybook/react-vite',
+    options: {},
+  },
+  docs: {
+    autodocs: 'tag',
+  },
+  viteFinal: config => {
+    return mergeConfig(config, {
+      resolve: {
+        alias: {
+          '@/*': resolve(__dirname, '../src'),
+        },
+      },
+    })
+  },
+}
+export default config
